Back up 4.6.

+++ src/Compiler/AST/Statement/Expression/IntConst.java
@@ -6,9 +6,9 @@ import static Compiler.Tool.Tool.indent;
  * Created by Alri on 16/3/31.
  */
 public class IntConst extends Expression {
-    public int intValue;
+    public long intValue;
 
-    public IntConst(int i) {
+    public IntConst(long i) {
         intValue = i;
     }
